+++ v3/crates/execute/src/plan/error.rs
@@ -1,4 +1,10 @@
-use open_dds::{relationships::RelationshipName, types::FieldName};
+use metadata_resolve::Qualified;
+use open_dds::{
+    arguments::ArgumentName,
+    commands::CommandName,
+    relationships::RelationshipName,
+    types::{CustomTypeName, FieldName},
+};
 use tracing_util::TraceableError;
 
 use crate::ndc;
@@ -23,6 +29,14 @@ pub enum InternalError {
     MappingExistsInRelationship {
         source_column: FieldName,
         relationship_name: RelationshipName,
+    },
+
+    #[error("Missing argument mapping to command {command_name} data connector source for argument {argument_name} used in relationship {relationship_name} on type {source_type}")]
+    MissingArgumentMappingInCommandRelationship {
+        source_type: Qualified<CustomTypeName>,
+        relationship_name: RelationshipName,
+        command_name: Qualified<CommandName>,
+        argument_name: ArgumentName,
     },
 
     #[error("remote relationships should have been handled separately")]